backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
alembic==1.13.1
asyncpg==0.29.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
plaid-python==14.0.0